--- a/addons/wfc/nodes/generator_2d.gd
+++ b/addons/wfc/nodes/generator_2d.gd
@@ -43,8 +43,6 @@ var render_intermediate_results: bool = false
 @export_category("Debug mode")
 var print_rules: bool = false
 
-var start_time: float
-
 signal done
 
 
@@ -90,7 +88,6 @@ func _exit_tree():
 		_runner = null
 
 func start():
-	start_time = Time.get_ticks_msec()
 	assert(_runner == null)
 	assert(target != null)
 	assert(rect.has_area())
@@ -117,7 +114,6 @@ func start():
 		if not rules.mapper.is_ready():
 			rules.mapper.learn_from(positive_sample_node)
 		var elapsed = (Time.get_ticks_msec() - start_learn);
-		print("Learn: %dms"%elapsed)
 		rules.learn_from(positive_sample_node)
 		
 		if rules.complete_matrices and negative_sample != null and not negative_sample.is_empty():
@@ -142,7 +138,7 @@ func start():
 	
 	_runner.start(problem)
 
-	_runner.all_solved.connect(func(): done.emit(); print(str(Time.get_ticks_msec() - start_time) + "ms"))
+	_runner.all_solved.connect(func(): done.emit())
 	_runner.sub_problem_solved.connect(_on_solved)
 	_runner.partial_solution.connect(_on_partial_solution)