--- a/addons/wfc/utils/layered_tile_data.gd
+++ b/addons/wfc/utils/layered_tile_data.gd
@@ -1,4 +1,0 @@
-extends Node
-
-var x: int
-var y: int