--- a/project.godot
+++ b/project.godot
@@ -11,15 +11,14 @@ config_version=5
 [application]
 
 config/name="WFC (Wave Function Collapse) for Godot 4"
-run/main_scene="res://addons/wfc/examples/demo_wfc_2d_tilemap.tscn"
-config/features=PackedStringArray("4.0", "GL Compatibility")
+run/main_scene="res://addons/wfc/examples/demo_wfc_2d_layered_tilemap.tscn"
+config/features=PackedStringArray("4.1", "GL Compatibility")
 config/icon="res://icon.png"
 
 [display]
 
 window/size/viewport_width=1920
 window/size/viewport_height=1080
-window/size/mode=3
 
 [editor]
 
@@ -31,5 +30,6 @@ enabled=PackedStringArray("res://addons/gut/plugin.cfg", "res://addons/wfc/plugi
 
 [rendering]
 
+textures/canvas_textures/default_texture_filter=0
 renderer/rendering_method="gl_compatibility"
 renderer/rendering_method.mobile="gl_compatibility"