… addons/wfc/examples/demo_wfc_2d_layered_tilemap.tscn