related to #2 

This is an implementation that adds a new `WFCLayeredTileMapMapper2D` class that will store cells as `Array[Vector4i]` instead of `Vector4i` to handle multiple layers.

It's working but it has many downsides:

 - It's really slower than the flat 2D approach
 - Not having weights makes the map go crazy quite easily (like say you have one dirt tile alone and one dirt + tree tile, you have 50% to get a tree, making it quite high)
 - Having to store these as `Array[Vector4i]` means that we're duplicating coords instead of just storing coords and then cell+alt in a better structure, that being said I'm not sure it'd make things much faster

This PR is mostly to open discussions on the implementation, it's clearly not ready for a release

**Performance**

Perf wise, the profiler shows 70ms avg per frame on my machine for the layered mapper vs 20ms for the flat one, this is way too much to be considered as a viable solution imho.

It appears that the # of entries in the sample is affecting performances quite hard, presumably because it creates more rules, but this shows a possible performance issue with cases that would be too complex.