@AlexeyBond 2023-11-07
> Not having weights makes the map go crazy quite easily (like say you have one dirt tile alone and one dirt + tree tile, you have 50% to get a tree, making it quite high)

That's expected behavior until tile probabilities (which are totaly independent issue) are implemented.

> Having to store these as `Array[Vector4i]` means that we're duplicating coords instead of just storing coords and then cell+alt in a better structure, that being said I'm not sure it'd make things much faster

Nothing is duplicated there. Coordinares received from `map.get_cell_atlas_coords(layer, coords)` are coordinates of a tile in texture atlas. `map.get_cell_source_id(layer, coords)` is index of a texture. So if there is only one texture and no alternative tiles in tile set, `map.get_cell_atlas_coords(layer, coords)` will be the only value that is different between a tree and dirt.

> It appears that the # of entries in the sample is affecting performances quite hard

I guess, a dictionary cannot work as a hash map when keys are mutable values (arrays) and thus it has to sequentially compare new value to all available keys. If that's the problem, I can suggest a possible solution

<details>
<summary>suggestion</summary>
I suggest to use a tree of dictionaries instead of single dictionary. E.g. instead of 

```
{
  [Vector4i(tile_from_layer_0), Vector4i(tile_from_layer1), ...]: 1
}
```

store

```
{
  Vector4i(tile_from_layer0): {
    Vector4i(tile_from_layer1): {
      ...: 1
    }
  }
}
```

That will make dictionary access functions more complicated:
- instead of `_read_cell_attrs` there will be two functions - one for `learn_from` that creates nested dictionaries when necessary and one for `read_cell` that returns `-1` when cannot find mapping on any nesting level.
- additional counter will be necessary to generate integer indices for tile types
- `_ensure_reverse_mapping` will have to generate arrays from keys in dictionary tree

But dictionary access time will become `O(average_tile_types_per_layer * layers)` (or even `O(layers)` if the dictionary will start working as a hash map) instead of `O(total_tile_combinations * layers)`.

</details>

However, I'm not sure if it's really a significant issue - `learn_from`/`read_cell` calls are not really frequent. What's more worth looking at is `write_cell`, which is called a lot if `render_intermediate_results` is enabled in generator. In fact, even `write_cell` isn't called that frequently when `render_intermediate_results` is disabled. And it is most likely to be disabled everywhere excapt this addon's examples. So, maybe performance of a mapper doesn't really matter that much.

> Perf wise, the profiler shows 70ms avg per frame on my machine for the layered mapper vs 20ms for the flat one, this is way too much to be considered as a viable solution imho.

Did you measure it with the same single-layer map, same sample (and same rules)? If the multi-layer mapper did actually work with more than one layer or rules were different then these measurements are not representative.