@Supamiu 2023-11-07
Yeah I realized that nothing was duplicated and forgot to remove that part of my message 😄 

So I made another performance test by recording start time and logging milliseconds duration when `done` signal emits, here are the results:

**WFCLayeredTileMapMapper2D**:
- With intermediate rendering: 14.2s
- Without intermediate rendering: 13.1s

**WFCTileMapMapper2D**
- With intermediate rendering: 10.7s
- Without intermediate rendering: 9.6s

*Tests made using the exact same setup, just swapped the mapper in the generator*

So turns out ms per frame isn't really representative here, I guess measuring like that is better 😄 . It shows that the most expensive part is not `write_cell` neither, I checked the learning phase and it's <1ms too so my guess is that the solver is the slow part here (which totally makes sense, we're introducing more possibilities and thus more rules).

I think your suggestion regarding using a 2 dimensions dictionary can help but I don't think I can implement that for now, my knowledge is too limited on the topic to make sure I'm not doing something bad.

The possible improvements I see are:
- Have the solver be more performant, if that's even possible, I didn't check it at all
- Have a way to lazy generate a map (like, render a given rect and trigger generator to generate more on demand instead of generating the whole map at once)

To elaborate more on the second one, we're generating a 121*68 map here, which is 8228 tiles in total, while the typical screen of a tilemap-based game would contain way more in its viewport (a rough check on my game, I'm around 20 * 12) which we could render as a priority part and then render the rest either while the player is moving or just in background as they start exploring the area.