@AlexeyBond 2023-11-07
I just tested both multi-layer and single-layer mappers on a single-layer tilemap (from original tilemap demo). This way all the other variables (including different rules) are excluded. And, suddenly multi-layer mapper works even a tiny bit faster than single-layer - generation takes 22.26 seconds instead of 22.38 seconds (average of 4 samples in each case).

So, I guess, it's ok to drop a single-layer mapper in favour of new multi-layer one.

I do have couple of comments for this PR and after resolving them it can be merged.

> Have the solver be more performant, if that's even possible, I didn't check it at all

I have spent some time optimizing it, so I guess it's current performance is about 80-90% of what is possible to acheive without dramatically changing the algorithm and/or technology.

> Have a way to lazy generate a map (like, render a given rect and trigger generator to generate more on demand instead of generating the whole map at once)

I think this can be done using some sort of higher-level node that spawns generators when player approaches ungenerated area. But it's not trivial. The generators should generate some extra area to ensure that there may be no contradictions in still unpopulated regions. Just like with multithreading. Also, it all just won't work for some rule sets and there is yet no correct way implemented to recognize such rules automatically.

One approach to this problem is to fill the whole map with some sort of "ground" tiles and then erase tiles in blocks before generating them, as described [here](https://www.boristhebrave.com/2021/11/08/infinite-modifying-in-blocks/). But it's likely to leave visible block borders in some rule sets.